# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApiBileMo",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Api for Phones Catalog in B2B"
        }
    ],
    "paths": {
        "/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "operationId": "App\\Controller\\CustomerController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "/customers?page=2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste de clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomersList"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun client pour cet utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun client pour cet utilisateur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerEdit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Création d'un client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le client a été ajouté avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "operationId": "App\\Controller\\CustomerController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "operationId": "App\\Controller\\CustomerController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerEdit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modification des informations client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le client a été modifié avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "operationId": "App\\Controller\\CustomerController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supression d'un client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le client a été supprimé avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomersList": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "self": {
                                    "type": "string"
                                },
                                "create": {
                                    "type": "string"
                                },
                                "update": {
                                    "type": "string"
                                },
                                "delete": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "embedded": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "phones": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "adress": {
                        "type": "string"
                    },
                    "postCode": {
                        "type": "integer"
                    },
                    "city": {
                        "type": "string"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "self": {
                                    "type": "string"
                                },
                                "create": {
                                    "type": "string"
                                },
                                "update": {
                                    "type": "string"
                                },
                                "delete": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "embedded": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "phones": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "CustomerEdit": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "adress": {
                        "type": "string"
                    },
                    "postCode": {
                        "type": "integer"
                    },
                    "city": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Phone": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "self": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "NotFound": {
                "description": "La ressource n'existe pas.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "La ressource n'existe pas."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID de la ressource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}